+++ common/src/main/java/com/teamresourceful/resourcefulconfig/api/types/entries/ResourcefulConfigObjectEntry.java
@@ -13,6 +13,10 @@ public interface ResourcefulConfigObjectEntry extends ResourcefulConfigEntry {
         return List.of();
     }
 
+    default Object instance() {
+        return null;
+    }
+
     default Component getTitle(@NotNull Component fallback) {
         return fallback;
     }