+++ common/src/main/java/com/teamresourceful/resourcefulconfig/api/types/entries/SerializableObject.java
@@ -1,0 +1,10 @@
+package com.teamresourceful.resourcefulconfig.api.types.entries;
+
+import com.google.gson.JsonElement;
+
+public interface SerializableObject {
+
+    JsonElement save();
+
+    void load(JsonElement json);
+}